Move `MCP::Annotations` to its own file
Also, implement `Annotations#to_h`
