+++ lib/mcp.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative "mcp/annotations"
 require_relative "mcp/configuration"
 require_relative "mcp/content"
 require_relative "mcp/instrumentation"
@@ -30,15 +31,6 @@ module MCP
 
     def configuration
       @configuration ||= Configuration.new
-    end
-  end
-
-  class Annotations
-    attr_reader :audience, :priority
-
-    def initialize(audience: nil, priority: nil)
-      @audience = audience
-      @priority = priority
     end
   end
 end
